7219 redo all the things camel-case files/folders again update tests & data to new structure
