+++ src/_scss/pages/dataDictionary/_dictionaryTable.scss
