+++ src/_scss/pages/dataDictionary/_searchBar.scss
