+++ src/js/apis/dataDictionary.js
@@ -1,0 +1,5 @@
+import { apiRequest } from '../helpers/apiRequest';
+
+export const fetchDataDictionary = () => apiRequest({
+  url: 'v2/references/data_dictionary/'
+});
