+++ src/js/containers/router/RouterRoutes.js
@@ -34,7 +34,7 @@ const AboutTheDataPage = React.lazy(() => import('components/aboutTheData/AboutT
 const AgencyDetailsPage = React.lazy(() => import('components/aboutTheData/AgencyDetailsPage').then((comp) => comp));
 const ErrorPage = React.lazy(() => import('components/errorPage/ErrorPage').then((comp) => comp));
 const SubmissionStatisticsDataSources = React.lazy(() => import('components/aboutTheData/DataSourcesAndMethodologiesPage').then((comp) => comp));
-const DataDictionaryPage = React.lazy(() => import('components/data-dictionary/data-dictionary-page').then((comp) => comp));
+const DataDictionaryPage = React.lazy(() => import('components/dataDictionary/dataDictionaryPage').then((comp) => comp));
 
 // /* eslint-disable import/prefer-default-export */
 // Please add any new routes to the scripts/pages.js routes file.
